+++ sdk/@launchdarkly/observability-python/ldobserve/_util/dict.py
@@ -1,9 +1,9 @@
-def _flatten_dict(d, parent_key="", sep="."):
+def flatten_dict(d, parent_key="", sep="."):
     items = []
     for k, v in d.items():
         new_key = parent_key + sep + k if parent_key else k
         if isinstance(v, dict):
-            items.extend(_flatten_dict(v, new_key, sep=sep).items())
+            items.extend(flatten_dict(v, new_key, sep=sep).items())
         else:
             items.append((new_key, v))
     return dict(items)
