--- sdk/@launchdarkly/observability-python/ldobserve/otel/__init__.py
@@ -1,23 +1,0 @@
-"""OpenTelemetry integration for LaunchDarkly Observability."""
-
-from .sampling_trace_exporter import (
-    SamplingTraceExporter,
-    ExportSampler,
-    _sample_spans,
-)
-from .sampling import (
-    CustomSampler,
-    SamplingResult,
-    default_sampler,
-)
-from .configuration import _OTELConfiguration
-
-__all__ = [
-    "SamplingTraceExporter",
-    "CustomSampler",
-    "SamplingResult",
-    "ExportSampler",
-    "_sample_spans",
-    "default_sampler",
-    "_OTELConfiguration",
-]
